fix: provided better feedback for upload errors
As far as I could investigate, we are using react-dropzone as the underlying upload component.
When passing `accept` or `maxSize` to `useDropzone`, it would simply ignore files that would
not comply with the requirements and therefore the validation would complain that the field
was empty, which is very confusing for the user.

The workaround was to lift the restrictions imposed by react-dropzone and manually validate
the requirements to display a proper error message.